requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
imbalanced-learn==0.11.0
joblib==1.3.2
Flask==3.0.0
python-dotenv==1.0.0
tqdm==4.66.1
streamlit==1.28.1
plotly==5.17.0
